refactor :: dto에 major 적용
