+++ src/main/java/com/example/daedongv3_5/domain/recruitment/presentation/dto/response/RecruitmentResponse.java
@@ -1,6 +1,6 @@
 package com.example.daedongv3_5.domain.recruitment.presentation.dto.response;
 
-import com.example.daedongv3_5.domain.club.domain.Major;
+import com.example.daedongv3_5.domain.club.domain.enums.Major;
 import com.example.daedongv3_5.domain.recruitment.domain.RecruitmentStatus;
 import lombok.Builder;
